requirements.txt
Flask==2.3.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.3
email-validator==1.3.1
pytest==7.4.0
Jinja2==3.1.2 
Werkzeug==2.3.7
python-dotenv==1.1.1
flask-admin==1.6.1